Update Clip datasets and changed raster output to float 32

+++ Hazard_Estimates/Raster_Sets.py
@@ -111,5 +111,5 @@ class raster_sets:
 
         predictions.return_dataset_2d(self.rasters[0].nrows)
         gc.collect()
-        self.rasters[0].save_image(predictions.asciiFile,location, file)
+        self.rasters[0].save_image(predictions.asciiFile.astype("float32"),location, file)
         return predictions
